feat(protocol-designer): decouple hardware and labware in deckSetup
closes AUTH-1716
