+++ protocol-designer/src/pages/Designer/DeckSetup/DeckSetupContainer.tsx
@@ -35,7 +35,10 @@ import { FixedTrashText } from '../../../components/molecules'
 import { getSelectedTerminalItemId } from '../../../ui/steps'
 import { selectors } from '../../../labware-ingred/selectors'
 import { DeckSetupDetails } from './DeckSetupDetails'
-import { DECK_SETUP_TOOLS_WIDTH_REM, DeckSetupTools } from './DeckSetupTools'
+import {
+  DECK_SETUP_TOOLS_WIDTH_REM,
+  DeckSetupToolbox,
+} from './DeckSetupToolbox'
 import {
   animateZoom,
   getCutoutIdForAddressableArea,
@@ -48,11 +51,9 @@ import type { StagingAreaLocation, TrashCutoutId } from '@opentrons/components'
 import type {
   AddressableAreaName,
   CutoutId,
-  ModuleModel,
   RobotType,
 } from '@opentrons/shared-data'
 import type { AdditionalEquipmentEntity } from '@opentrons/step-generation'
-import type { Fixture } from './constants'
 
 const WASTE_CHUTE_SPACE = 30
 const DETAILS_HOVER_SPACE = 60
@@ -140,8 +141,6 @@ export function DeckSetupContainer(
   }, '')
 
   const [hoveredLabware, setHoveredLabware] = useState<string | null>(null)
-  const [hoveredModule, setHoveredModule] = useState<ModuleModel | null>(null)
-  const [hoveredFixture, setHoveredFixture] = useState<Fixture | null>(null)
 
   const addEquipment = (slotId: string): void => {
     const cutoutId =
@@ -344,8 +343,6 @@ export function DeckSetupContainer(
                   <DeckSetupDetails
                     selectedZoomInSlot={zoomIn.slot ?? undefined}
                     hoveredLabware={hoveredLabware}
-                    hoveredModule={hoveredModule}
-                    hoveredFixture={hoveredFixture}
                     hover={hoverSlot}
                     terminalItemId={terminalItemId}
                     setHover={setHoverSlot}
@@ -369,11 +366,7 @@ export function DeckSetupContainer(
           </Flex>
         </Flex>
         {zoomIn.slot != null && zoomIn.cutout != null ? (
-          <DeckSetupTools
-            onDeckProps={{
-              setHoveredFixture,
-              setHoveredModule,
-            }}
+          <DeckSetupToolbox
             onCloseClick={() => {
               dispatch(selectZoomedIntoSlot({ slot: null, cutout: null }))
               animateZoom({
