+++ protocol-designer/src/pages/Designer/DeckSetup/DeckSetupDetails.tsx
@@ -51,7 +51,6 @@ import type {
   CutoutId,
   DeckDefinition,
   DeckSlotId,
-  ModuleModel,
 } from '@opentrons/shared-data'
 import type {
   InitialDeckSetup,
@@ -59,16 +58,13 @@ import type {
   ModuleOnDeck,
 } from '../../../step-forms'
 import type { DeckSetupTerminalIdType } from '../types'
-import type { Fixture } from './constants'
 
 interface DeckSetupDetailsProps extends DeckSetupTerminalIdType {
   activeDeckSetup: InitialDeckSetup
   addEquipment: (slotId: string) => void
   deckDef: DeckDefinition
   hover: string | null
-  hoveredFixture: Fixture | null
   hoveredLabware: string | null
-  hoveredModule: ModuleModel | null
   setHover: Dispatch<SetStateAction<string | null>>
   showGen1MultichannelCollisionWarnings: boolean
   stagingAreaCutoutIds: CutoutId[]
@@ -81,9 +77,7 @@ export function DeckSetupDetails(props: DeckSetupDetailsProps): JSX.Element {
     addEquipment,
     deckDef,
     hover,
-    hoveredFixture,
     hoveredLabware: hoveredLabwareFromProp,
-    hoveredModule,
     selectedZoomInSlot,
     terminalItemId,
     setHover,
@@ -556,20 +550,14 @@ export function DeckSetupDetails(props: DeckSetupDetailsProps): JSX.Element {
       <SelectedHoveredItems
         deckDef={deckDef}
         robotType={robotType}
-        hoveredFixture={hoveredFixture}
         hoveredLabware={hoveredLabwareFromProp}
-        hoveredModule={hoveredModule}
         slotPosition={slotPosition}
       />
 
-      {/* hovered hardware + labware */}
+      {/* hovered  labware */}
       <HoveredItems
         hoveredSlotPosition={slotPosition}
-        deckDef={deckDef}
-        robotType={robotType}
-        hoveredFixture={hoveredFixture}
         hoveredLabware={hoveredLabwareFromProp}
-        hoveredModule={hoveredModule}
       />
 
       {/* slot overflow menu */}
