+++ protocol-designer/src/pages/Designer/DeckSetup/SlotOverflowMenu.tsx
@@ -15,12 +15,7 @@ import {
   StyledText,
   useOnClickOutside,
 } from '@opentrons/components'
-import {
-  FLEX_ROBOT_TYPE,
-  FLEX_STAGING_AREA_SLOT_ADDRESSABLE_AREAS,
-  getCutoutIdFromAddressableArea,
-  getDeckDefFromRobotType,
-} from '@opentrons/shared-data'
+import { FLEX_STAGING_AREA_SLOT_ADDRESSABLE_AREAS } from '@opentrons/shared-data'
 
 import { getRobotType } from '../../../file-data/selectors'
 import {
@@ -29,7 +24,6 @@ import {
   openIngredientSelector,
 } from '../../../labware-ingred/actions'
 import { getNextAvailableDeckSlot } from '../../../labware-ingred/utils'
-import { deleteModule } from '../../../modules'
 import {
   ConfirmDeleteEntityInUseModal,
   ConfirmDeleteStagingAreaModal,
@@ -37,21 +31,17 @@ import {
 } from '../../../components/organisms'
 import { getSavedStepForms } from '../../../step-forms/selectors'
 import { useKitchen } from '../../../components/organisms/Kitchen/hooks'
-import { deleteDeckFixture } from '../../../step-forms/actions/additionalItems'
 import { getDeckSetupForActiveItem } from '../../../top-selectors/labware-locations'
-import { getStagingAreaAddressableAreas } from '../../../utils'
 import { selectors as labwareIngredSelectors } from '../../../labware-ingred/selectors'
-import { getIsEntityOnSlotInUse } from './utils'
+import { getIsLabwareOnSlotInUse } from './utils'
 
 import type { MouseEvent, SetStateAction } from 'react'
 import type {
   AddressableAreaName,
   CoordinateTuple,
-  CutoutId,
   DeckSlotId,
 } from '@opentrons/shared-data'
 
-import type { LabwareOnDeck } from '../../../step-forms'
 import type { ThunkDispatch } from '../../../types'
 
 const ROBOT_BOTTOM_HALF_SLOTS = [
@@ -127,11 +117,7 @@ export function SlotOverflowMenu(
 
   const { makeSnackbar } = useKitchen()
 
-  const {
-    labware: deckSetupLabware,
-    modules: deckSetupModules,
-    additionalEquipmentOnDeck,
-  } = deckSetup
+  const { labware: deckSetupLabware, modules: deckSetupModules } = deckSetup
 
   const isOffDeckLocation = deckSetupLabware[location] != null
 
@@ -158,27 +144,7 @@ export function SlotOverflowMenu(
   const nestedLabwareOnSlot = Object.values(deckSetupLabware).find(
     lw => lw.slot === labwareOnSlot?.id
   )
-
-  const fixturesOnSlot = Object.values(additionalEquipmentOnDeck).filter(
-    ae => ae.location?.split('cutout')[1] === location
-  )
-  const stagingAreaCutout = fixturesOnSlot.find(
-    fixture => fixture.name === 'stagingArea'
-  )?.location
-
-  let matchingLabware: LabwareOnDeck | null = null
-  if (stagingAreaCutout != null) {
-    const stagingAreaAddressableAreaName = getStagingAreaAddressableAreas([
-      stagingAreaCutout,
-    ] as CutoutId[])
-    matchingLabware =
-      Object.values(deckSetupLabware).find(
-        lw => lw.slot === stagingAreaAddressableAreaName[0]
-      ) ?? null
-  }
-
-  const hasNoItems =
-    moduleOnSlot == null && labwareOnSlot == null && fixturesOnSlot.length === 0
+  const hasNoItems = labwareOnSlot == null && nestedLabwareOnSlot == null
 
   const isStagingSlot = FLEX_STAGING_AREA_SLOT_ADDRESSABLE_AREAS.includes(
     location as AddressableAreaName
@@ -202,24 +168,13 @@ export function SlotOverflowMenu(
     setShowMenuList(false)
   }
 
-  const isEntityOnSlotInUse = getIsEntityOnSlotInUse(
+  const isLabwareOnSlotInUse = getIsLabwareOnSlotInUse(
     savedSteps,
-    matchingLabware,
-    moduleOnSlot,
     labwareOnSlot,
-    nestedLabwareOnSlot,
-    fixturesOnSlot
+    nestedLabwareOnSlot
   )
 
   const handleClear = (): void => {
-    //  clear module from slot
-    if (moduleOnSlot != null) {
-      dispatch(deleteModule({ moduleId: moduleOnSlot.id }))
-    }
-    //  clear fixture(s) from slot
-    if (fixturesOnSlot.length > 0) {
-      fixturesOnSlot.forEach(fixture => dispatch(deleteDeckFixture(fixture.id)))
-    }
     //  clear labware from slot
     if (labwareOnSlot != null) {
       dispatch(deleteContainer({ labwareId: labwareOnSlot.id }))
@@ -227,25 +182,6 @@ export function SlotOverflowMenu(
     //  clear nested labware from slot
     if (nestedLabwareOnSlot != null) {
       dispatch(deleteContainer({ labwareId: nestedLabwareOnSlot.id }))
-    }
-    // clear labware on staging area 4th column slot
-    if (matchingLabware != null) {
-      dispatch(deleteContainer({ labwareId: matchingLabware.id }))
-    }
-    // delete staging slot if addressable area is on staging slot
-    if (isStagingSlot) {
-      const deckDef = getDeckDefFromRobotType(FLEX_ROBOT_TYPE)
-      const cutoutId = getCutoutIdFromAddressableArea(location, deckDef)
-      const stagingAreaEquipmentId = Object.values(
-        additionalEquipmentOnDeck
-      ).find(({ location }) => location === cutoutId)?.id
-      if (stagingAreaEquipmentId != null) {
-        dispatch(deleteDeckFixture(stagingAreaEquipmentId))
-      } else {
-        console.error(
-          `could not find equipment id for entity in ${location} with cutout id ${cutoutId}`
-        )
-      }
     }
   }
 
@@ -287,17 +223,12 @@ export function SlotOverflowMenu(
 
   const slotOverflowBody = (
     <>
-      {isEntityOnSlotInUse && showDeleteEntityInUseModal ? (
+      {isLabwareOnSlotInUse && showDeleteEntityInUseModal ? (
         <ConfirmDeleteEntityInUseModal
           onConfirm={() => {
-            if (matchingLabware != null) {
-              setShowDeleteLabwareModal(true)
-              setShowDeleteEntityInUseModal(false)
-            } else {
-              handleClear()
-              setShowMenuList(false)
-              setShowDeleteEntityInUseModal(false)
-            }
+            handleClear()
+            setShowMenuList(false)
+            setShowDeleteEntityInUseModal(false)
           }}
           onClose={() => {
             setShowDeleteEntityInUseModal(false)
@@ -346,9 +277,7 @@ export function SlotOverflowMenu(
           }}
         >
           <StyledText desktopStyle="bodyDefaultRegular">
-            {hasNoItems
-              ? t(isOffDeckLocation ? 'add_labware' : 'add_hw_lw')
-              : t(isOffDeckLocation ? 'edit_labware' : 'edit_hw_lw')}
+            {hasNoItems ? t('add_labware') : t('edit_labware')}
           </StyledText>
         </MenuItem>
         {canRenameLabwareAndEditLiquids ? (
@@ -390,12 +319,8 @@ export function SlotOverflowMenu(
         <MenuItem
           disabled={hasNoItems && !isStagingSlot}
           onClick={(e: MouseEvent) => {
-            if (isEntityOnSlotInUse) {
+            if (isLabwareOnSlotInUse) {
               setShowDeleteEntityInUseModal(true)
-              e.preventDefault()
-              e.stopPropagation()
-            } else if (!isEntityOnSlotInUse && matchingLabware != null) {
-              setShowDeleteLabwareModal(true)
               e.preventDefault()
               e.stopPropagation()
             } else {
@@ -405,7 +330,7 @@ export function SlotOverflowMenu(
           }}
         >
           <StyledText desktopStyle="bodyDefaultRegular">
-            {t(isOffDeckLocation ? 'clear_labware' : 'clear_slot')}
+            {t('clear_labware')}
           </StyledText>
         </MenuItem>
       </Flex>
