+++ protocol-designer/src/pages/Designer/DeckSetup/__tests__/DeckSetupContainer.test.tsx
@@ -16,14 +16,14 @@ import { getSelectedTerminalItemId } from '../../../../ui/steps'
 import { getDisableModuleRestrictions } from '../../../../feature-flags/selectors'
 import { getRobotType } from '../../../../file-data/selectors'
 import { DeckSetupDetails } from '../DeckSetupDetails'
-import { DeckSetupTools } from '../DeckSetupTools'
+import { DeckSetupToolbox } from '../DeckSetupToolbox'
 import { DeckSetupContainer } from '../DeckSetupContainer'
 import type * as OpentronsComponents from '@opentrons/components'
 
 vi.mock('../../../../ui/steps/selectors')
 vi.mock('../../../../top-selectors/labware-locations')
 vi.mock('../../../../feature-flags/selectors')
-vi.mock('../DeckSetupTools')
+vi.mock('../DeckSetupToolbox')
 vi.mock('../DeckSetupDetails')
 vi.mock('../../../../ui/steps')
 vi.mock('../../../../labware-ingred/selectors')
@@ -56,7 +56,9 @@ describe('DeckSetupContainer', () => {
       { id: null, text: null },
     ])
     vi.mocked(getHoveredDropdownItem).mockReturnValue({ id: null, text: null })
-    vi.mocked(DeckSetupTools).mockReturnValue(<div>mock DeckSetupTools</div>)
+    vi.mocked(DeckSetupToolbox).mockReturnValue(
+      <div>mock DeckSetupToolbox</div>
+    )
     vi.mocked(DeckSetupDetails).mockReturnValue(
       <div>mock DeckSetupDetails</div>
     )
@@ -71,12 +73,12 @@ describe('DeckSetupContainer', () => {
       pipettes: {},
     })
   })
-  it('renders the decksetupTools when slot and cutout are not null', () => {
+  it('renders the DeckSetupToolbox when slot and cutout are not null', () => {
     render()
     screen.getByText('mock DeckSetupDetails')
-    screen.getByText('mock DeckSetupTools')
+    screen.getByText('mock DeckSetupToolbox')
   })
-  it('renders no deckSetupTools when slot and cutout are null', () => {
+  it('renders no DeckSetupToolbox when slot and cutout are null', () => {
     vi.mocked(selectors.getZoomedInSlot).mockReturnValue({
       slot: null,
       cutout: null,
