+++ protocol-designer/src/pages/Designer/DeckSetup/__tests__/DeckSetupToolbox.test.tsx
@@ -1,28 +1,20 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import '@testing-library/jest-dom/vitest'
 import { fireEvent, screen } from '@testing-library/react'
-import {
-  ABSORBANCE_READER_V1,
-  FLEX_ROBOT_TYPE,
-  HEATERSHAKER_MODULE_V1,
-  fixture96Plate,
-} from '@opentrons/shared-data'
-import { GRIPPER_LOCATION } from '@opentrons/step-generation'
+import { FLEX_ROBOT_TYPE, fixture96Plate } from '@opentrons/shared-data'
 import { i18n } from '../../../../assets/localization'
 import { renderWithProviders } from '../../../../__testing-utils__'
 import { deleteContainer } from '../../../../labware-ingred/actions'
 import { useKitchen } from '../../../../components/organisms/Kitchen/hooks'
-import { deleteModule } from '../../../../modules'
 import {
   getAdditionalEquipment,
   getSavedStepForms,
 } from '../../../../step-forms/selectors'
 import { getRobotType } from '../../../../file-data/selectors'
-import { deleteDeckFixture } from '../../../../step-forms/actions/additionalItems'
 import { selectors } from '../../../../labware-ingred/selectors'
 import { getDismissedHints } from '../../../../tutorial/selectors'
 import { getDeckSetupForActiveItem } from '../../../../top-selectors/labware-locations'
-import { DeckSetupTools } from '../DeckSetupTools'
+import { DeckSetupToolbox } from '../DeckSetupToolbox'
 import { LabwareTools } from '../LabwareTools'
 
 import type { ComponentProps } from 'react'
@@ -33,31 +25,25 @@ vi.mock('../../../../feature-flags/selectors')
 vi.mock('../../../../file-data/selectors')
 vi.mock('../../../../top-selectors/labware-locations')
 vi.mock('../../../../labware-ingred/actions')
-vi.mock('../../../../modules')
-vi.mock('../../../../step-forms/actions/additionalItems')
 vi.mock('../../../../labware-ingred/selectors')
 vi.mock('../../../../tutorial/selectors')
 vi.mock('../../../../step-forms/selectors')
 vi.mock('../../../../components/organisms/Kitchen/hooks')
-const render = (props: ComponentProps<typeof DeckSetupTools>) => {
-  return renderWithProviders(<DeckSetupTools {...props} />, {
+const render = (props: ComponentProps<typeof DeckSetupToolbox>) => {
+  return renderWithProviders(<DeckSetupToolbox {...props} />, {
     i18nInstance: i18n,
   })[0]
 }
 
 const mockMakeSnackbar = vi.fn()
 
-describe('DeckSetupTools', () => {
-  let props: ComponentProps<typeof DeckSetupTools>
+describe('DeckSetupToolbox', () => {
+  let props: ComponentProps<typeof DeckSetupToolbox>
 
   beforeEach(() => {
     props = {
       onCloseClick: vi.fn(),
       setHoveredLabware: vi.fn(),
-      onDeckProps: {
-        setHoveredModule: vi.fn(),
-        setHoveredFixture: vi.fn(),
-      },
     }
     vi.mocked(selectors.getZoomedInSlotInfo).mockReturnValue({
       selectedLabwareDefUri: null,
@@ -86,29 +72,9 @@ describe('DeckSetupTools', () => {
   afterEach(() => {
     vi.resetAllMocks()
   })
-  it('should render the relevant modules and fixtures for slot D3 on Flex with tabs', () => {
+  it('should render the labware', () => {
     render(props)
-    screen.getByText('Add a module')
-    screen.getByText('Add a fixture')
-    screen.getByTestId('DeckInfoLabel_D3')
     screen.getByText('Customize slot')
-    screen.getByText('Deck hardware')
-    screen.getByText('Labware')
-    screen.getByText('Absorbance Plate Reader Module GEN1')
-    screen.getByText('Heater-Shaker Module GEN1')
-    screen.getByText('Magnetic Block GEN1')
-    screen.getByText('Temperature Module GEN2')
-    screen.getByText('Staging area')
-    screen.getByText('Waste Chute')
-    screen.getByText('Trash Bin')
-    screen.getByText('Waste Chute with Staging Area')
-    screen.getByText('Magnetic Block GEN1 with Staging Area')
-  })
-  it('should render the labware tab', () => {
-    render(props)
-    screen.getByText('Deck hardware')
-    // click on labware tab
-    fireEvent.click(screen.getByText('Labware'))
     screen.getByText('mock labware tools')
   })
   it('should clear the slot from all items when the clear cta is called', () => {
@@ -155,52 +121,5 @@ describe('DeckSetupTools', () => {
     render(props)
     fireEvent.click(screen.getByText('Clear'))
     expect(vi.mocked(deleteContainer)).toHaveBeenCalledTimes(2)
-    expect(vi.mocked(deleteModule)).toHaveBeenCalled()
-    expect(vi.mocked(deleteDeckFixture)).toHaveBeenCalled()
-  })
-  it('should close and preserve h-s module when done is called', () => {
-    vi.mocked(selectors.getZoomedInSlotInfo).mockReturnValue({
-      selectedLabwareDefUri: null,
-      selectedNestedLabwareDefUri: null,
-      selectedFixture: null,
-      selectedModuleModel: HEATERSHAKER_MODULE_V1,
-      selectedSlot: { slot: 'D3', cutout: 'cutoutD3' },
-    })
-    render(props)
-    fireEvent.click(screen.getByText('Heater-Shaker Module GEN1'))
-    fireEvent.click(screen.getByText('Done'))
-    expect(props.onCloseClick).toHaveBeenCalled()
-  })
-  it('should close and preserve waste chute and staging area when done is called', () => {
-    vi.mocked(selectors.getZoomedInSlotInfo).mockReturnValue({
-      selectedLabwareDefUri: null,
-      selectedNestedLabwareDefUri: null,
-      selectedFixture: 'wasteChuteAndStagingArea',
-      selectedModuleModel: HEATERSHAKER_MODULE_V1,
-      selectedSlot: { slot: 'D3', cutout: 'cutoutD3' },
-    })
-    render(props)
-    fireEvent.click(screen.getByText('Waste Chute with Staging Area'))
-    fireEvent.click(screen.getByText('Done'))
-    expect(props.onCloseClick).toHaveBeenCalled()
-  })
-  it('should save plate reader if gripper configured', () => {
-    vi.mocked(getAdditionalEquipment).mockReturnValue({
-      gripperUri: {
-        name: 'gripper',
-        id: 'gripperId',
-        location: GRIPPER_LOCATION,
-      },
-    })
-    vi.mocked(selectors.getZoomedInSlotInfo).mockReturnValue({
-      selectedLabwareDefUri: null,
-      selectedNestedLabwareDefUri: null,
-      selectedFixture: null,
-      selectedModuleModel: ABSORBANCE_READER_V1,
-      selectedSlot: { slot: 'D3', cutout: 'cutoutD3' },
-    })
-    render(props)
-    fireEvent.click(screen.getByText('Done'))
-    expect(props.onCloseClick).toHaveBeenCalled()
   })
 })
