+++ protocol-designer/src/pages/Designer/DeckSetup/__tests__/HoveredItems.test.tsx
@@ -2,33 +2,23 @@ import { describe, it, vi, beforeEach } from 'vitest'
 import '@testing-library/jest-dom/vitest'
 import { screen } from '@testing-library/react'
 import { renderWithProviders } from '../../../../__testing-utils__'
-import {
-  FLEX_ROBOT_TYPE,
-  TEMPERATURE_MODULE_V2,
-  getDeckDefFromRobotType,
-} from '@opentrons/shared-data'
-import { LabwareRender, Module } from '@opentrons/components'
+import { LabwareRender } from '@opentrons/components'
 import { selectors } from '../../../../labware-ingred/selectors'
 import { getCustomLabwareDefsByURI } from '../../../../labware-defs/selectors'
 import { getSelectedTerminalItemId } from '../../../../ui/steps'
-import { FixtureRender } from '../FixtureRender'
 import { HoveredItems } from '../HoveredItems'
-import { ModuleLabel } from '../ModuleLabel'
 
 import type { ComponentProps } from 'react'
 import type * as OpentronsComponents from '@opentrons/components'
 
-vi.mock('../FixtureRender')
 vi.mock('../../../../labware-ingred/selectors')
 vi.mock('../../../../labware-defs/selectors')
-vi.mock('../ModuleLabel')
 vi.mock('../../../../ui/steps')
 vi.mock('@opentrons/components', async importOriginal => {
   const actual = await importOriginal<typeof OpentronsComponents>()
   return {
     ...actual,
     LabwareRender: vi.fn(),
-    Module: vi.fn(),
   }
 })
 
@@ -42,11 +32,7 @@ describe('HoveredItems', () => {
   beforeEach(() => {
     props = {
       hoveredSlotPosition: [0, 0, 0],
-      deckDef: getDeckDefFromRobotType(FLEX_ROBOT_TYPE),
-      robotType: FLEX_ROBOT_TYPE,
       hoveredLabware: null,
-      hoveredModule: null,
-      hoveredFixture: 'trashBin',
     }
     vi.mocked(selectors.getZoomedInSlotInfo).mockReturnValue({
       selectedLabwareDefUri: null,
@@ -56,28 +42,13 @@ describe('HoveredItems', () => {
       selectedSlot: { slot: 'D3', cutout: 'cutoutD3' },
     })
     vi.mocked(getCustomLabwareDefsByURI).mockReturnValue({})
-    vi.mocked(FixtureRender).mockReturnValue(<div>mock FixtureRender</div>)
     vi.mocked(LabwareRender).mockReturnValue(<div>mock LabwareRender</div>)
-    vi.mocked(Module).mockReturnValue(<div>mock Module</div>)
-    vi.mocked(ModuleLabel).mockReturnValue(<div>mock ModuleLabel</div>)
     vi.mocked(getSelectedTerminalItemId).mockReturnValue('__initial_setup__')
   })
-  it('renders a hovered fixture', () => {
-    render(props)
-    screen.getByText('mock FixtureRender')
-  })
   it('renders a hovered labware', () => {
-    props.hoveredFixture = null
     props.hoveredLabware = 'fixture/fixture_universal_flat_bottom_adapter/1'
     render(props)
     screen.getByText('mock LabwareRender')
     screen.getByText('Fixture Opentrons Universal Flat Heater-Shaker Adapter')
-  })
-  it('renders a hovered module', () => {
-    props.hoveredFixture = null
-    props.hoveredModule = TEMPERATURE_MODULE_V2
-    render(props)
-    screen.getByText('mock Module')
-    screen.getByText('mock ModuleLabel')
   })
 })
